# docs/contract/openapi.yaml
openapi: 3.1.0
info:
  title: Bardic server API
  version: 0.2.6
  summary: Contract between the Bardic server and its clients (draft).
  description: |
    Normative, language-neutral contract for Bardic v2. It is derived from
    `docs/PRODUCT-SPEC.md`, not from the prototype. Clients are generated
    from this file; the server must satisfy it. Read `README.md` in this folder
    first: it explains identity headers, errors, money, events, long-running
    work and the place-sync protocol once, so each operation does not repeat them.

    Status: pre-release (0.x). Until 1.0.0, additive changes bump the patch
    version and breaking changes bump the minor version.
  license:
    name: Personal use
    identifier: LicenseRef-Personal
servers:
  - url: http://localhost:8765
    description: Default local address. The server name and address are configurable.
...
paths:
...
  /api/books/{book_id}/deletion:
    parameters: [{$ref: '#/components/parameters/BookId'}]
    get:
      operationId: getBookDeletion
      tags: [Admin]
      summary: Is this book scheduled for permanent deletion
      description: 404 `deletion_not_found` when none is scheduled.
      x-bardic-cost: none
      responses:
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '500': {$ref: '#/components/responses/ServerError'}
        '200':
          description: The schedule.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Deletion'}
        '404': {$ref: '#/components/responses/NotFound'}
    post:
      operationId: scheduleBookDeletion
      tags: [Admin]
      summary: Permanently delete a book in 60 seconds
      description: |
        Hides the book immediately and schedules irreversible deletion of the
        book, its audio, places, history and plans `delay_seconds` later (default
        and minimum 60, at most 7 days). The schedule is stored and survives
        restarts. Cancel with `cancelBookDeletion`. Audited. The spending ledger
        and the audit log are kept. Errors (409): `job_running` (audio is being
        made for the book), `deletion_pending`, `book_adding`.
      x-bardic-cost: none
      parameters: [{$ref: '#/components/parameters/Listener'}, {$ref: '#/components/parameters/Device'}]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                delay_seconds: {type: integer, minimum: 60, default: 60}
      responses:
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '500': {$ref: '#/components/responses/ServerError'}
        '202':
          description: Scheduled.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Deletion'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
    delete:
      operationId: cancelBookDeletion
      tags: [Admin]
      summary: Undo a scheduled deletion
      description: The book returns exactly as it was. 409 `deletion_done` once it ran.
      x-bardic-cost: none
      parameters: [{$ref: '#/components/parameters/Listener'}, {$ref: '#/components/parameters/Device'}]
      responses:
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '500': {$ref: '#/components/responses/ServerError'}
        '204': {description: "Cancelled."}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
...
  /api/backups:
    get:
      operationId: listBackups
      tags: [Admin]
      summary: Backups made by this server
      x-bardic-cost: none
      responses:
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '500': {$ref: '#/components/responses/ServerError'}
        '200':
          description: Backups.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: {$ref: '#/components/schemas/Backup'}
    post:
      operationId: createBackup
      tags: [Admin]
      summary: Make a consistent backup while running
      description: |
        Writes a point-in-time copy of the database (one consistent snapshot) to
        `backups/<id>/bardic.db` in the data folder, and links the original books
        and the finished audio into `backups/<id>/media/`. Media files are
        immutable, so they are hard links (a copy where the filesystem cannot
        link), not second copies. The backup is a folder on the same computer: copy
        it elsewhere to survive a lost disk. Returns a job-like `Backup`; poll
        `listBackups` for `done`.
      x-bardic-cost: none
      parameters: [{$ref: '#/components/parameters/Device'}]
      responses:
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '500': {$ref: '#/components/responses/ServerError'}
        '202':
          description: Started.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Backup'}
...
components:
  parameters:
    Listener:
      name: X-Bardic-Listener
      in: header
      required: true
      description: Id of the listener acting. Scopes places and settings; unknown id is 404 `listener_not_found`.
      schema: {type: string}
...
    Device:
      name: X-Bardic-Device
      in: header
      required: true
      description: Client-generated device id (UUID). Registers the device on first use and is recorded in the audit log.
      schema: {type: string}
...
    BookId: {name: book_id, in: path, required: true, schema: {type: string}}
...
  responses:
    InvalidRequest:
      description: The request is malformed or breaks a rule; `code` says which.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    NotFound:
      description: The resource does not exist (or is not visible to this listener).
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    Conflict:
      description: The request conflicts with current state; `code` says how.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    ServerError:
      description: An unexpected failure on the server. Nothing you need to undo was changed; `retryable` is true.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
...
  schemas:
...
    Error:
      type: object
      required: [code, detail]
      properties:
        code: {type: string, description: "Stable machine-readable code (open enumeration); see README."}
        detail: {type: string, description: "Human-readable, safe to show."}
        retryable: {type: boolean}
        retry_after_seconds: {type: [integer, 'null']}
        context:
          type: object
          additionalProperties: true
          description: Code-specific fields (for example `plan_id`, `existing_book_id`).
...
    Actor:
      type: object
      required: [listener_id, listener_name, device_id, device_name]
      description: Who acted. The listener is null when no listener existed or applied (for example renaming the server, or creating the first listener).
      properties:
        listener_id: {type: [string, 'null']}
        listener_name: {type: [string, 'null']}
        device_id: {type: string}
        device_name: {type: string}
...
    Deletion:
      type: object
      required: [book_id, scheduled_at, executes_at, state]
      properties:
        book_id: {type: string}
        scheduled_at: {type: string, format: date-time}
        executes_at: {type: string, format: date-time}
        state: {type: string, enum: [pending, cancelled, done]}
        scheduled_by: {$ref: '#/components/schemas/Actor'}
...
    Backup:
      type: object
      required: [id, state, created_at, bytes, path]
      properties:
        id: {type: string}
        state: {type: string, enum: [running, done, failed]}
        created_at: {type: string, format: date-time}
        bytes: {type: [integer, 'null']}
        path: {type: [string, 'null'], description: "Location on the Bardic computer."}
...